% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivph_laplace}
\alias{bivph_laplace}
\title{Bivariate phase-type joint Laplace}
\usage{
bivph_laplace(r, alpha, S11, S12, S22)
}
\arguments{
\item{r}{matrix of values}

\item{alpha}{vector of initial probabilities}

\item{S11}{sub-intensity matrix}

\item{S12}{matrix}

\item{S22}{sub-intensity matrix}
}
\value{
Joint laplace at \code{r}
}
\description{
Bivariate phase-type joint Laplace
}
\examples{
alpha <- c(0.15, 0.85)
S11 <- matrix(c(c(-2, 9), c(0, -11)), nrow = 2, ncol = 2)
S12 <- matrix(c(c(2, 0), c(0, 2)), nrow = 2, ncol = 2)
S22 <- matrix(c(c(-1, 0), c(0.5, -5)), nrow = 2, ncol = 2)
x <- matrix(c(c(0.5, 1), c(2, 1.5)), ncol = 2)
bivph_laplace(x, alpha, S11, S12, S22)
}
