% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bm_dens_aux_cov}
\alias{bm_dens_aux_cov}
\title{Correlated Phase-type mixing with covariates}
\usage{
bm_dens_aux_cov(n, ex1, w, ex2, alpha, S11, S12, S22)
}
\arguments{
\item{n}{vector of frequencies.}

\item{w}{vector of aggregated severities.}

\item{alpha}{Vector of initial probabilities.}

\item{S11}{Sub-intensity matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-intensity matrix.}
}
\value{
The density at \code{x}.
}
\description{
Computes the joint density of a correlated phase-type Mixed Poisson distribution
with parameters \code{alpha} and \code{S} at \code{x}.
}
