% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{cdf,frailty-method}
\alias{cdf,frailty-method}
\title{Distribution method for univariate phase type frailty models}
\usage{
\S4method{cdf}{frailty}(x, q, X = numeric(0), lower.tail = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{frailty}.}

\item{q}{A vector of locations.}

\item{X}{A matrix of covariates.}

\item{lower.tail}{Logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A vector containing the corresponding CDF evaluations.
}
\description{
Distribution method for univariate phase type frailty models
}
\examples{
obj <- frailty(phasetype(structure = "general"), bhaz = "weibull", bhaz_pars = 2)
cdf(obj, c(1, 2, 3))
}
