% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{cdf,mp-method}
\alias{cdf,mp-method}
\title{Distribution method for univariate phase type mixed Poisson models}
\usage{
\S4method{cdf}{mp}(x, q, X = numeric(0), lower.tail = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{mp}.}

\item{q}{A vector of locations.}

\item{X}{A matrix of covariates.}

\item{lower.tail}{Logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A vector containing the CDF evaluations at the given locations.
}
\description{
Distribution method for univariate phase type mixed Poisson models
}
\examples{
obj <- mp(phasetype(structure = "general"))
cdf(obj, c(1, 2, 3))
}
