% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{cdf,phasetype-method}
\alias{cdf,phasetype-method}
\title{Distribution method for phase type distributions}
\usage{
\S4method{cdf}{phasetype}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{phasetype}.}

\item{q}{a vector of locations.}

\item{lower.tail}{logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A list containing the locations and corresponding CDF evaluations.
}
\description{
Distribution method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
cdf(obj, c(1, 2, 3))
}
