% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{coef,bivphasetype-method}
\alias{coef,bivphasetype-method}
\title{Coef method for bivariate phasetype class}
\usage{
\S4method{coef}{bivphasetype}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivphasetype}.}
}
\value{
Parameters of bivariate phase type model.
}
\description{
Coef method for bivariate phasetype class
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
coef(obj)
}
