% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{coef,bphasetype-method}
\alias{coef,bphasetype-method}
\title{Coef method for bivariate phasetype class}
\usage{
\S4method{coef}{bphasetype}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bphasetype}.}
}
\value{
Parameters of bivariate phase type model.
}
\description{
Coef method for bivariate phasetype class
}
\examples{
obj <- bphasetype(dimensions = c(3,3))
coef(obj)
}
