% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{coef,correlated-method}
\alias{coef,correlated-method}
\title{Coef method for correlated frailty class}
\usage{
\S4method{coef}{correlated}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{correlated}.}
}
\value{
Parameters of the correlated phase-type frailty model.
}
\description{
Coef method for correlated frailty class
}
\examples{
bivph_obj <- bivphasetype(dimensions = c(3, 3))
cobj <- correlated(bivph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
coef(cobj)
}
