% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{coef,frailty-method}
\alias{coef,frailty-method}
\title{Coef Method for frailty Class}
\usage{
\S4method{coef}{frailty}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{frailty}.}
}
\value{
parameters of frailty model.
}
\description{
Coef Method for frailty Class
}
\examples{
obj <- frailty(phasetype(structure = "general", dimension = 2), gfun = "lognormal", gfun_pars = 2)
coef(obj)
}
