% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{coef,mp-method}
\alias{coef,mp-method}
\title{Coef method for univariate phase-type mixed Poisson class}
\usage{
\S4method{coef}{mp}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{mp}.}
}
\value{
Parameters of the univariate phase type mixed Poisson model.
}
\description{
Coef method for univariate phase-type mixed Poisson class
}
\examples{
obj <- mp(phasetype(structure = "general", dimension = 2))
coef(obj)
}
