% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{coef,phasetype-method}
\alias{coef,phasetype-method}
\title{Coef Method for phasetype Class}
\usage{
\S4method{coef}{phasetype}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{phasetype}.}
}
\value{
Parameters of phasetype model.
}
\description{
Coef Method for phasetype Class
}
\examples{
obj <- phasetype(structure = "general")
coef(obj)
}
