% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{coef,phfrailty-method}
\alias{coef,phfrailty-method}
\title{Coef Method for phfrailty Class}
\usage{
\S4method{coef}{phfrailty}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{iph}.}
}
\value{
parameters of phfrailty model.
}
\description{
Coef Method for phfrailty Class
}
\examples{
obj <- phfrailty(ph(structure = "general", dimension = 2), gfun = "lognormal", gfun_pars = 2)
coef(obj)
}
