% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{coef,shared-method}
\alias{coef,shared-method}
\title{Coef method for shared frailty class}
\usage{
\S4method{coef}{shared}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{shared}.}
}
\value{
Parameters of the shared phase type frailty model.
}
\description{
Coef method for shared frailty class
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
coef(shared_obj)
}
