% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{corr,bivphasetype-method}
\alias{corr,bivphasetype-method}
\title{Correlation method for bivphasetype class}
\usage{
\S4method{corr}{bivphasetype}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivphasetype}.}

\item{k}{A vector with the location.}
}
\value{
An real value.
}
\description{
Correlation method for bivphasetype class
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
corr(obj)
}
