% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.mph.R
\name{corr,mph-method}
\alias{corr,mph-method}
\title{Correlation method for mph class}
\usage{
\S4method{corr}{mph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{k}{A vector with the location.}
}
\value{
An real value.
}
\description{
Correlation method for mph class
}
\examples{
obj <- mph(structure = c("general", "general"))
corr(obj)
}
