% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{corr}
\alias{corr}
\title{New generic for the correlation of multivariate matrix distributions}
\usage{
corr(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Correlation of the multivariate object
}
\description{
Methods are available for objects of classes \linkS4class{phasetype} and
\linkS4class{bivphasetype}.
}
