% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\docType{class}
\name{correlated-class}
\alias{correlated-class}
\title{Correlated phase-type frailty model}
\value{
Class object.
}
\description{
Class of objects for correlated phase-type frailty models.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the bivariate phase-type distribution.}

\item{\code{bhaz1}}{A list comprising of the parameters.}

\item{\code{bhaz2}}{A list comprising of the parameters.}

\item{\code{coefs}}{Regression parameters.}
}}

