% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{correlated}
\alias{correlated}
\title{Constructor function for shared phase type frailty models}
\usage{
correlated(
  bivph = NULL,
  bhaz1 = NULL,
  bhaz_pars1 = NULL,
  bhaz2 = NULL,
  bhaz_pars2 = NULL,
  B = numeric(0),
  alpha = NULL,
  S11 = NULL,
  S12 = NULL,
  S22 = NULL,
  dimensions = c(3, 3)
)
}
\arguments{
\item{bivph}{An object of class \linkS4class{bivphasetype}.}

\item{bhaz1}{Baseline hazard function of first marginal.}

\item{bhaz_pars1}{The parameters of the baseline hazard function of first marginal.}

\item{bhaz2}{Baseline hazard function of first marginal.}

\item{bhaz_pars2}{The parameters of the baseline hazard function of first marginal.}

\item{B}{Regression parameters.}

\item{alpha}{A probability vector.}

\item{S11}{A sub-intensity matrix.}

\item{S12}{A matrix.}

\item{S22}{A sub-intensity matrix.}

\item{dimensions}{The dimension of the bivariate phase-type structure (if provided).}
}
\value{
An object of class \linkS4class{correlated}.
}
\description{
Constructor function for shared phase type frailty models
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
correlated(obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
}
