% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{cross,shared-method}
\alias{cross,shared-method}
\title{Cross method for shared phase-type frailty models}
\usage{
\S4method{cross}{shared}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{shared}.}

\item{y}{A matrix of locations.}
}
\value{
A vector containing the evaluations of the cross-ratio function for a
shared phase-type frailty model at the given locations.
}
\description{
Cross method for shared phase-type frailty models
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
cross(shared_obj, matrix(c(1, 2), ncol = 2))
}
