% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{cross}
\alias{cross}
\title{New generic for the cross ratio function of matrix distributions}
\usage{
cross(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Cross ratio function from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{phasetype}.
}
