% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{dens,bivphasetype-method}
\alias{dens,bivphasetype-method}
\title{Density method for bivariate phase type distributions}
\usage{
\S4method{dens}{bivphasetype}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{bivphasetype}.}

\item{y}{A matrix of locations.}
}
\value{
A vector containing the joint density evaluations at the given locations.
}
\description{
Density method for bivariate phase type distributions
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
dens(obj, matrix(c(0.5, 1), ncol = 2))
}
