% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{dens,bphasetype-method}
\alias{dens,bphasetype-method}
\title{Density method for bivariate phase type distributions}
\usage{
\S4method{dens}{bphasetype}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{bphasetype}.}

\item{y}{A matrix of locations.}
}
\value{
A list containing the locations and corresponding joint density evaluations.
}
\description{
Density method for bivariate phase type distributions
}
\examples{
obj <- bphasetype(dimensions = c(3,3))
dens(obj, matrix(c(0.5, 1), ncol = 2))
}
