% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{dens,correlated-method}
\alias{dens,correlated-method}
\title{Density method for correlated phase type frailty models}
\usage{
\S4method{dens}{correlated}(x, y, X = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{correlated}.}

\item{y}{A matrix of locations.}

\item{X}{A matrix of covariates.}
}
\value{
A vector containing the joint density evaluations at the given locations.
}
\description{
Density method for correlated phase type frailty models
}
\examples{
bivph_obj <- bivphasetype(dimensions = c(3,3))
cobj <- correlated(bivph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
dens(cobj, matrix(c(1, 2), ncol = 2))
}
