% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{dens,frailty-method}
\alias{dens,frailty-method}
\title{Density method for univariate phase type frailty models}
\usage{
\S4method{dens}{frailty}(x, y, X = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{frailty}.}

\item{y}{A vector of locations.}

\item{X}{A matrix of covariates.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for univariate phase type frailty models
}
\examples{
obj <- frailty(phasetype(structure = "general"), bhaz = "weibull", bhaz_pars = 2)
dens(obj, c(1, 2, 3))
}
