% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{dens,mp-method}
\alias{dens,mp-method}
\title{Density method for univariate phase type mixed Poisson models}
\usage{
\S4method{dens}{mp}(x, y, X = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{mp}.}

\item{y}{A vector of locations.}

\item{X}{A matrix of covariates.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for univariate phase type mixed Poisson models
}
\examples{
obj <- mp(phasetype(structure = "general"))
dens(obj, c(1, 2, 3))
}
