% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{dens,phasetype-method}
\alias{dens,phasetype-method}
\title{Density method for phase type distributions}
\usage{
\S4method{dens}{phasetype}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{y}{A vector of locations.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
dens(obj, c(1, 2, 3))
}
