% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{dens,phfrailty-method}
\alias{dens,phfrailty-method}
\title{Density Method for inhomogeneous phase type distributions}
\usage{
\S4method{dens}{phfrailty}(x, y)
}
\arguments{
\item{x}{an object of class \linkS4class{phfrailty}.}

\item{y}{a vector of locations.}
}
\value{
A list containing the locations and corresponding density evaluations.
}
\description{
Density Method for inhomogeneous phase type distributions
}
\examples{
obj <- phfrailty(ph(structure = "general"), gfun = "weibull", gfun_pars = 2)
dens(obj, c(1, 2, 3))
}
