% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{dens}
\alias{dens}
\title{New generic for the density of matrix distributions}
\usage{
dens(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
Density from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{phasetype}
}
