% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretizate_density}
\alias{discretizate_density}
\title{Discretization of a univariate density}
\usage{
discretizate_density(
  density,
  parameters,
  ini_point,
  truncation_point,
  max_deltat,
  max_probability
)
}
\arguments{
\item{density}{The density function.}

\item{parameters}{Parameters of the density function.}

\item{ini_point}{Initial value for the discretization.}

\item{truncation_point}{Max value for the discretization.}

\item{max_deltat}{Maximum size of interval.}

\item{max_probability}{Maximum probability of a discrete point.}
}
\value{
List with values and weights.
}
\description{
Discretizates a univariate density function using Simpson's rule.
}
