% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_bivph}
\alias{EMstep_bivph}
\title{EM  for bivariate PH using Matlab algorithm for matrix exponential}
\usage{
EMstep_bivph(alpha, S11, S12, S22, obs, weight)
}
\arguments{
\item{alpha}{initial probabilities}

\item{S11}{sub-intensity}

\item{S12}{a matrix}

\item{S22}{sub-intensity}

\item{obs}{the observations}

\item{weight}{the weights for the observations}
}
\value{
fitted alpha, S11, S12 and S22 after one iteration
}
\description{
EM  for bivariate PH using Matlab algorithm for matrix exponential
}
