% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_PADE}
\alias{EMstep_PADE}
\title{EM using Matlab algorithm for matrix exponential in combination with Armadillo}
\usage{
EMstep_PADE(alpha, S, obs, weight)
}
\arguments{
\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{the weights for the observations}
}
\value{
fitted alpha and S after one iteration
}
\description{
EM using Matlab algorithm for matrix exponential in combination with Armadillo
}
