% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{fit,correlated-method}
\alias{fit,correlated-method}
\title{Fit method for correlated phase type frailty models}
\usage{
\S4method{fit}{correlated}(
  x,
  y,
  X = numeric(0),
  stepsEM = 100,
  stepsPH = 50,
  initialpoint1 = 1e-04,
  truncationpoint1 = 8,
  delta1 = 0.1,
  initialpoint2 = 1e-04,
  truncationpoint2 = 8,
  delta2 = 0.1,
  maxit = 100,
  reltol = 1e-08,
  every = 100
)
}
\arguments{
\item{x}{An object of class \linkS4class{correlated}.}

\item{y}{Matrix or data.}

\item{X}{A matrix of covariates.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{stepsPH}{Number of EM steps for the phase-type component at each iteration of the global EM.}

\item{initialpoint1}{Initial value for discretization of first marginal density.}

\item{truncationpoint1}{Ultimate value for discretization of first marginal density.}

\item{delta1}{Size of interval for discretization of first marginal.}

\item{initialpoint2}{Initial value for discretization of second marginal density.}

\item{truncationpoint2}{Ultimate value for discretization of second marginal density.}

\item{delta2}{Size of interval for discretization of second marginal.}

\item{maxit}{Maximum number of iterations when optimizing g function.}

\item{reltol}{Relative tolerance when optimizing g function.}

\item{every}{Number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{correlated}.
}
\description{
Fit method for correlated phase type frailty models
}
\examples{
bivph_obj <- bivphasetype(dimensions = c(3, 3))
cobj <- correlated(bivph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
data <- sim(cobj, n = 100)
fit(cobj, data, stepsEM = 5, stepsPH = 5, every = 1)
}
