% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{fit,mp-method}
\alias{fit,mp-method}
\title{Fit method for phase-type mixed Poisson models}
\usage{
\S4method{fit}{mp}(
  x,
  y,
  exposure = numeric(0),
  weight = numeric(0),
  X = numeric(0),
  stepsEM = 100,
  stepsPH = 50,
  initialpoint = 1e-04,
  truncationpoint = 10,
  maxprobability = 0.01,
  maxdelta = 0.05,
  maxit = 100,
  reltol = 1e-08,
  every = 100
)
}
\arguments{
\item{x}{An object of class \linkS4class{mp}.}

\item{y}{Vector or data.}

\item{exposure}{Vector of exposures for observations.}

\item{weight}{Vector of weights for observations.}

\item{X}{A matrix of covariates.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{stepsPH}{Number of EM steps for the phase-type component at each iteration of the global EM.}

\item{initialpoint}{Initial value for discretization of density.}

\item{truncationpoint}{Ultimate value for discretization of density.}

\item{maxprobability}{Max probability allowed for an interval in the discretization.}

\item{maxdelta}{Max size of interval allowed for the discretization.}

\item{maxit}{Maximum number of iterations when optimizing g function.}

\item{reltol}{Relative tolerance when optimizing g function.}

\item{every}{Number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{mp}.
}
\description{
Fit method for phase-type mixed Poisson models
}
\examples{
obj <- mp(phasetype(structure = "general", dimension = 2))
data <- sim(obj, n = 100)
fit(obj, data, stepsEM = 50, every = 10)
}
