% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{fit,phasetype-method}
\alias{fit,phasetype-method}
\title{Fit method for phasetype Class}
\usage{
\S4method{fit}{phasetype}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  stepsEM = 1000,
  methods = c("RK", "RK"),
  rkstep = NA,
  uni_epsilon = NA,
  maxit = 100,
  reltol = 1e-08,
  every = 100,
  plot = FALSE
)
}
\arguments{
\item{x}{an object of class \linkS4class{phasetype}.}

\item{y}{vector or data.}

\item{weight}{vector of weights.}

\item{rcen}{vector of right-censored observations}

\item{rcenweight}{vector of weights for right-censored observations.}

\item{stepsEM}{number of EM steps to be performed.}

\item{methods}{methods to use for matrix exponential calculation: RM, UNI or PADE}

\item{rkstep}{Runge-Kutta step size (optional)}

\item{uni_epsilon}{epsilon parameter for uniformization method}

\item{maxit}{maximum number of iterations when optimizing g function.}

\item{reltol}{relative tolerance when optimizing g function.}

\item{every}{number of iterations between likelihood display updates.}

\item{plot}{logical indicating whether to plot the fit at each iteration.}
}
\value{
An object of class \linkS4class{phasetype}.
}
\description{
Fit method for phasetype Class
}
\examples{
obj <- frailty(phasetype(structure = "general", dimension = 2), bhaz = "weibull", bhaz_pars = 2)
data <- sim(obj, n = 100)
fit(obj, data, stepsEM = 1000, every = 200)
}
