% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{fit,shared-method}
\alias{fit,shared-method}
\title{Fit method for shared phase type frailty models}
\usage{
\S4method{fit}{shared}(
  x,
  y,
  rcen = numeric(0),
  X = numeric(0),
  stepsEM = 1000,
  stepsPH = 50,
  initialpoint = 1e-04,
  truncationpoint = 10,
  maxprobability = 0.01,
  maxdelta = 0.05,
  maxit = 100,
  reltol = 1e-08,
  every = 100
)
}
\arguments{
\item{x}{An object of class \linkS4class{shared}.}

\item{y}{Matrix or data.}

\item{rcen}{Matrix of indicators of right-censored observations}

\item{X}{A matrix of covariates.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{stepsPH}{Number of EM steps for the phase-type component at each iteration of the global EM.}

\item{initialpoint}{Initial value for discretization of density.}

\item{truncationpoint}{Ultimate value for discretization of density.}

\item{maxprobability}{Max probability allowed for an interval in the discretization.}

\item{maxdelta}{Max size of interval allowed for the discretization.}

\item{maxit}{Maximum number of iterations when optimizing g function.}

\item{reltol}{Relative tolerance when optimizing g function.}

\item{every}{Number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{shared}.
}
\description{
Fit method for shared phase type frailty models
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
data <- sim(shared_obj, n = 100)
fit(shared_obj, data, stepsEM = 50, every = 10)
}
