% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\docType{class}
\name{frailty-class}
\alias{frailty-class}
\title{Phase type frailty model}
\value{
Class object.
}
\description{
Class of objects for univariate phase type frailty models.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{bhaz}}{A list comprising of the parameters.}

\item{\code{coefs}}{Regression parameters.}
}}

