% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{frailty}
\alias{frailty}
\title{Constructor Function for phase type frailty models}
\usage{
frailty(
  ph = NULL,
  gfun = NULL,
  gfun_pars = NULL,
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3,
  scale = 1
)
}
\arguments{
\item{ph}{An object of class \linkS4class{phasetype}.}

\item{gfun}{baseline hazard function}

\item{gfun_pars}{the parameters of the baseline hazard function}

\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid phase-type structure}

\item{dimension}{the dimension of the phase-type structure (if provided)}

\item{scale}{scale}
}
\value{
An object of class \linkS4class{frailty}.
}
\description{
Constructor Function for phase type frailty models
}
\examples{
frailty(phasetype(structure = "coxian", dimension = 4), gfun = "pareto", gfun_pars = 3)
}
