% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{haz,phasetype-method}
\alias{haz,phasetype-method}
\title{Hazard rate method for phase type distributions}
\usage{
\S4method{haz}{phasetype}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{y}{A vector of locations.}
}
\value{
A vector containing the hazard rate evaluations at the given locations.
}
\description{
Hazard rate method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
haz(obj, c(1, 2, 3))
}
