% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_norm}
\alias{inf_norm}
\title{L_inf norm of a matrix}
\usage{
inf_norm(A)
}
\arguments{
\item{A}{A matrix.}
}
\value{
The inf norm of A.
}
\description{
Computes the inf norm of a matrix A, defined as
L_inf A =  max (1 <= I <= M) sum (1 <= J <= N) abs (A(I,J)).
}
