% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{logLik,phasetype-method}
\alias{logLik,phasetype-method}
\title{logLik method for phasetype class}
\usage{
\S4method{logLik}{phasetype}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{phasetype}.}
}
\value{
An object of class logLik.
}
\description{
logLik method for phasetype class
}
\examples{
obj <- frailty(phasetype(structure = "general", dimension = 2), bhaz = "weibull", bhaz_pars = 2)
data <- sim(obj, n = 100)
fitted_ph <- fit(obj, data, stepsEM = 10)
logLik(fitted_ph)
}
