% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{marginal,bivphasetype-method}
\alias{marginal,bivphasetype-method}
\title{Marginal method for bivphasetype class}
\usage{
\S4method{marginal}{bivphasetype}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{bivphasetype}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{phasetype}.
}
\description{
Marginal method for bivphasetype class
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
marginal(obj, 1)
}
