% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{marginal,correlated-method}
\alias{marginal,correlated-method}
\title{Marginal method for correlated frailty class}
\usage{
\S4method{marginal}{correlated}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{correlated}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{frailty}.
}
\description{
Marginal method for correlated frailty class
}
\examples{
bivph_obj <- bivphasetype(dimensions = c(3, 3))
cobj <- correlated(bivph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
marginal(cobj, 1)
}
