% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{marginal,shared-method}
\alias{marginal,shared-method}
\title{Marginal method for shared frailty class}
\usage{
\S4method{marginal}{shared}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{shared}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{frailty}.
}
\description{
Marginal method for shared frailty class
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
marginal(shared_obj, 1)
}
