% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\docType{class}
\name{mp-class}
\alias{mp-class}
\title{Phase type mixed Poisson model}
\value{
Class object.
}
\description{
Class of objects for univariate phase type mixed Poisson models.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{coefs}}{Regression parameters.}
}}

