% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{mp}
\alias{mp}
\title{Constructor function for univariate phase type mixed Poisson models}
\usage{
mp(
  ph = NULL,
  B = numeric(0),
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3
)
}
\arguments{
\item{ph}{An object of class \linkS4class{phasetype}.}

\item{B}{Regression parameters.}

\item{alpha}{A probability vector.}

\item{S}{A sub-intensity matrix.}

\item{structure}{A valid phase-type structure.}

\item{dimension}{The dimension of the phase-type structure (if provided).}
}
\value{
An object of class \linkS4class{mp}.
}
\description{
Constructor function for univariate phase type mixed Poisson models
}
\examples{
mp(phasetype(structure = "coxian", dimension = 4))
}
