% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mp_aux_density}
\alias{mp_aux_density}
\title{Phase-type mixed Poisson density with covariates}
\usage{
mp_aux_density(x, ex, alpha, S)
}
\arguments{
\item{x}{Non-negative integer values.}

\item{ex}{Vector of covariates effect.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of a phase-type Mixed Poisson distribution with parameters
\code{alpha} and \code{S} at \code{x}.
}
