% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mp_cor_dens_aux}
\alias{mp_cor_dens_aux}
\title{Correlated Phase-type mixed Poisson density with covariates}
\usage{
mp_cor_dens_aux(x, ex, alpha, S11, S12, S22)
}
\arguments{
\item{x}{Matrix of values.}

\item{ex}{Matrix of covariates effect.}

\item{alpha}{Vector of initial probabilities.}

\item{S11}{Sub-intensity matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-intensity matrix.}
}
\value{
The density at \code{x}.
}
\description{
Computes the joint density of a correlated phase-type Mixed Poisson distribution
with parameters \code{alpha} and \code{S} at \code{x}.
}
