% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_cdf}
\alias{ph_cdf}
\title{Phase-type cdf or tail}
\usage{
ph_cdf(x, alpha, S, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative values}

\item{alpha}{vector of initial probabilities}

\item{S}{sub-intensity matrix}

\item{lower_tail}{cdf or tail}
}
\value{
The density at \code{x}
}
\description{
Computes the cdf or tail of a phase-type distribution with parameters \code{alpha} and \code{S} at \code{x}
}
