% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_laplace}
\alias{ph_laplace}
\title{Laplace transform of a phase-type distribution}
\usage{
ph_laplace(r, alpha, S)
}
\arguments{
\item{r}{real value}

\item{alpha}{vector of initial probabilities}

\item{S}{sub-intensity matrix}
}
\value{
Laplace transform at \code{r}
}
\description{
Computes the Laplace transform at \code{r} of a phase-type distribution with parameters \code{alpha} and \code{S}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
S <- matrix(c(c(-1, 0, 0), c(1, -2, 0),c(0, 1, -5)), nrow = 3, ncol = 3)
ph_laplace(0.5, alpha, S)
}
