% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{phasetype}
\alias{phasetype}
\title{Constructor function for phase type distributions}
\usage{
phasetype(alpha = NULL, S = NULL, structure = NULL, dimension = 3)
}
\arguments{
\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid phase-type structure ("general", "coxian", "hyperexponential", "gcoxian", "gerlang").}

\item{dimension}{the dimension of the phase-type structure (if structure is provided).}
}
\value{
An object of class \linkS4class{phasetype}.
}
\description{
Constructor function for phase type distributions
}
\examples{
phasetype(structure = "gcoxian", dim = 5)
phasetype(alpha = c(.5, .5), S = matrix(c(-1, .5, .5, -1), 2, 2))
}
