% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmph}
\alias{rmph}
\title{Simulate a MPH* random vector}
\usage{
rmph(n, alpha, S, R)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward matrix.}
}
\value{
The simulated sample.
}
\description{
Generates a sample of size \code{n} from a MPH* distribution with parameters
\code{alpha}, \code{S} and \code{R}.
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
S <- matrix(c(c(-1, 0, 0), c(1, -2, 0), c(0, 1, -5)), nrow = 3, ncol = 3)
R <- matrix(c(c(1, 0, 0.8), c(0, 1, 0.2)), nrow = 3, ncol = 2)
n <- 10
rmph(n, alpha, S, R)
}
