% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rphasetype}
\alias{rphasetype}
\title{Random phase-type}
\usage{
rphasetype(n, alpha, S)
}
\arguments{
\item{n}{sample size}

\item{alpha}{vector of initial probabilities}

\item{S}{sub-intensity matrix}
}
\value{
simulated sample
}
\description{
Generates a sample of size \code{n} from a phase-type distribution with parameters \code{alpha} and \code{S}
}
