% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\docType{class}
\name{shared-class}
\alias{shared-class}
\title{Shared phase type frailty model}
\value{
Class object.
}
\description{
Class of objects for shared phase type frailty models.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{bhaz1}}{A list comprising of the parameters.}

\item{\code{bhaz2}}{A list comprising of the parameters.}

\item{\code{coefs}}{Regression parameters.}
}}

