% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{shared}
\alias{shared}
\title{Constructor function for shared phase type frailty models}
\usage{
shared(
  ph = NULL,
  bhaz1 = NULL,
  bhaz_pars1 = NULL,
  bhaz2 = NULL,
  bhaz_pars2 = NULL,
  B = numeric(0),
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3
)
}
\arguments{
\item{ph}{an object of class \linkS4class{phasetype}.}

\item{bhaz1}{baseline hazard function of first marginal}

\item{bhaz_pars1}{the parameters of the baseline hazard function of first marginal}

\item{bhaz2}{baseline hazard function of first marginal}

\item{bhaz_pars2}{the parameters of the baseline hazard function of first marginal}

\item{B}{regression parameters}

\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid phase-type structure}

\item{dimension}{the dimension of the phase-type structure (if provided)}
}
\value{
An object of class \linkS4class{shared}.
}
\description{
Constructor function for shared phase type frailty models
}
\examples{
obj <- phasetype(structure = "coxian")
shared(obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
}
