% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{show,bivphasetype-method}
\alias{show,bivphasetype-method}
\title{Show method for bivariate phase-type distributions}
\usage{
\S4method{show}{bivphasetype}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivphasetype}.}
}
\description{
Show method for bivariate phase-type distributions
}
