% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{show,bphasetype-method}
\alias{show,bphasetype-method}
\title{Show method for bivariate phase type distributions}
\usage{
\S4method{show}{bphasetype}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bphasetype}.}
}
\description{
Show method for bivariate phase type distributions
}
