% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.correlatedph.R
\name{show,correlated-method}
\alias{show,correlated-method}
\title{Show method for correlated phase type frailty models}
\usage{
\S4method{show}{correlated}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{correlated}.}
}
\description{
Show method for correlated phase type frailty models
}
