% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{show,frailty-method}
\alias{show,frailty-method}
\title{Show method for phase type frailty models}
\usage{
\S4method{show}{frailty}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{frailty}.}
}
\description{
Show method for phase type frailty models
}
