% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{show,mp-method}
\alias{show,mp-method}
\title{Show method for univariate phase type mixed Poisson models}
\usage{
\S4method{show}{mp}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{mp}.}
}
\description{
Show method for univariate phase type mixed Poisson models
}
