% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{show,phfrailty-method}
\alias{show,phfrailty-method}
\title{Show Method for inhomogeneous phase type distributions}
\usage{
\S4method{show}{phfrailty}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{iph}.}
}
\description{
Show Method for inhomogeneous phase type distributions
}
