% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{show,shared-method}
\alias{show,shared-method}
\title{Show method for shared phase type frailty models}
\usage{
\S4method{show}{shared}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{shared}.}
}
\description{
Show method for shared phase type frailty models
}
