% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{sim,bivphasetype-method}
\alias{sim,bivphasetype-method}
\title{Simulation method for bivariate phase-type distributions}
\usage{
\S4method{sim}{bivphasetype}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{bivphasetype}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed bivariate
phase-type vector.
}
\description{
Simulation method for bivariate phase-type distributions
}
\examples{
obj <- bivphasetype(dimensions = c(3, 3))
sim(obj, n = 100)
}
