% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bivph.R
\name{sim,bphasetype-method}
\alias{sim,bphasetype-method}
\title{Simulation method for bivariate phase type distributions}
\usage{
\S4method{sim}{bphasetype}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{bphasetype}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed phase-type variables.
}
\description{
Simulation method for bivariate phase type distributions
}
\examples{
obj <- bphasetype(dimensions = c(3,3))
sim(obj, n = 100)
}
