% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{sim,frailty-method}
\alias{sim,frailty-method}
\title{Simulation method for univariate phase type frailty models}
\usage{
\S4method{sim}{frailty}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{frailty}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed phase-type frailty variables.
}
\description{
Simulation method for univariate phase type frailty models
}
\examples{
obj <- frailty(phasetype(structure = "general"), bhaz = "weibull", bhaz_pars = 2)
sim(obj, n = 100)
}
