% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.mixpoi.R
\name{sim,mp-method}
\alias{sim,mp-method}
\title{Simulation method for univariate phase type mixed Poisson models}
\usage{
\S4method{sim}{mp}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{mp}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed phase-type mixed Poisson variables.
}
\description{
Simulation method for univariate phase type mixed Poisson models
}
\examples{
obj <- mp(phasetype(structure = "general"))
sim(obj, n = 100)
}
