% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{sim,phasetype-method}
\alias{sim,phasetype-method}
\title{Simulation method for phase type distributions}
\usage{
\S4method{sim}{phasetype}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed phase-type variables.
}
\description{
Simulation method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
sim(obj, n = 100)
}
