% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{sim,phfrailty-method}
\alias{sim,phfrailty-method}
\title{Simulation Method for inhomogeneous phase type distributions}
\usage{
\S4method{sim}{phfrailty}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{iph}.}

\item{n}{an integer of length of realization.}
}
\value{
A realization of independent and identically distributed inhomogeneous phase-type variables.
}
\description{
Simulation Method for inhomogeneous phase type distributions
}
\examples{
obj <- phfrailty(ph(structure = "general"), gfun = "lognormal", gfun_pars = 2)
sim(obj, n = 100)
}
