% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{sim,shared-method}
\alias{sim,shared-method}
\title{Simulation method for shared phase type frailty models}
\usage{
\S4method{sim}{shared}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{shared}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed shared phase-type frailty random vectors.
}
\description{
Simulation method for shared phase type frailty models
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
sim(shared_obj, n = 100)
}
