% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{surv,frailty-method}
\alias{surv,frailty-method}
\title{Survival method for univariate phase type frailty models}
\usage{
\S4method{surv}{frailty}(x, q, X = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{frailty}.}

\item{q}{A vector of locations.}

\item{X}{A matrix of covariates.}
}
\value{
A vector containing the survival function evaluations at the given locations.
}
\description{
Survival method for univariate phase type frailty models
}
\examples{
obj <- frailty(phasetype(structure = "general"), bhaz = "weibull", bhaz_pars = 2)
surv(obj, c(1, 2, 3))
}
