% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{surv,phasetype-method}
\alias{surv,phasetype-method}
\title{Survival method for phase-type distributions}
\usage{
\S4method{surv}{phasetype}(x, q)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{q}{A vector of locations.}
}
\value{
A vector containing the survival function evaluations at the given locations.
}
\description{
Survival method for phase-type distributions
}
\examples{
obj <- phasetype(structure = "general")
surv(obj, c(1, 2, 3))
}
