% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{surv}
\alias{surv}
\title{New generic for the survival function of matrix distributions}
\usage{
surv(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Survival function from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{phasetype}.
}
