% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices}
\alias{vector_of_matrices}
\title{Computes elements S^n until the value size}
\usage{
vector_of_matrices(theVector, S, sizevect)
}
\arguments{
\item{theVector}{a vector}

\item{S}{sub-intensity matrix}

\item{sizevect}{size of vector}
}
\description{
Computes elements S^n until the value size
}
