% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{cdf,frailty-method}
\alias{cdf,frailty-method}
\title{Distribution Method for inhomogeneous phase type distributions}
\usage{
\S4method{cdf}{frailty}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{frailty}.}

\item{q}{a vector of locations.}

\item{lower.tail}{logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A list containing the locations and corresponding CDF evaluations.
}
\description{
Distribution Method for inhomogeneous phase type distributions
}
