% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{cdf,phasetype-method}
\alias{cdf,phasetype-method}
\title{Distribution method for phase-type distributions}
\usage{
\S4method{cdf}{phasetype}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{q}{A vector of locations.}

\item{lower.tail}{Logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A vector containing the CDF (or tail) evaluations at the given locations.
}
\description{
Distribution method for phase-type distributions
}
\examples{
obj <- phasetype(structure = "general")
cdf(obj, c(1, 2, 3))
}
