% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{coef,frailty-method}
\alias{coef,frailty-method}
\title{Coef method for univariate frailty class}
\usage{
\S4method{coef}{frailty}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{frailty}.}
}
\value{
parameters of univariate phase type frailty model.
}
\description{
Coef method for univariate frailty class
}
\examples{
obj <- frailty(phasetype(structure = "general", dimension = 2), bhaz = "weibull", bhaz_pars = 2)
coef(obj)
}
