% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{coef,phasetype-method}
\alias{coef,phasetype-method}
\title{Coef method for phasetype class}
\usage{
\S4method{coef}{phasetype}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{phasetype}.}
}
\value{
Parameters of phase type model.
}
\description{
Coef method for phasetype class
}
\examples{
obj <- phasetype(structure = "general")
coef(obj)
}
