% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretizate_bivdensity}
\alias{discretizate_bivdensity}
\title{Discretization of an bivariate density}
\usage{
discretizate_bivdensity(
  density,
  parameters,
  ini_point1,
  truncation_point1,
  max_deltat1,
  ini_point2,
  truncation_point2,
  max_deltat2,
  max_probability
)
}
\arguments{
\item{density}{the density function}

\item{parameters}{parameters of the density function}

\item{ini_point1}{max value for the discretization - first component}

\item{truncation_point1}{max value for the discretization - first component}

\item{max_deltat1}{maximum size of interval - first component}

\item{ini_point2}{max value for the discretization - second component}

\item{truncation_point2}{max value for the discretization - second component}

\item{max_deltat2}{maximum size of interval - second component}

\item{max_probability}{maximum probability of a discrete point}
}
\value{
List with values and weights
}
\description{
Discretizates an bivariate density function using Simposon rule
}
