% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretizate_density}
\alias{discretizate_density}
\title{Discretization of a density}
\usage{
discretizate_density(
  density,
  parameters,
  truncation_point,
  max_probability,
  max_deltat
)
}
\arguments{
\item{density}{the density function}

\item{parameters}{parameters of the density function}

\item{truncation_point}{max value for the discretization}

\item{max_probability}{maximum probability of a discrete point}

\item{max_deltat}{maximum size of interval}
}
\value{
List with values and weights
}
\description{
Discretizates a density function using Simposon rule
}
