% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{fit,phasetype-method}
\alias{fit,phasetype-method}
\title{Fit method for phase-type frailty models}
\usage{
\S4method{fit}{phasetype}(
  x,
  y,
  rcen = numeric(0),
  stepsEM = 1000,
  stepsPH = 50,
  initialpoint = 1e-04,
  truncationpoint = 10,
  maxprobability = 0.01,
  maxdelta = 0.05,
  maxit = 100,
  reltol = 1e-08,
  every = 100
)
}
\arguments{
\item{x}{an object of class \linkS4class{phasetype}.}

\item{y}{vector or data.}

\item{rcen}{vector of right-censored observations}

\item{stepsEM}{number of EM steps to be performed.}

\item{stepsPH}{number of EM steps for the phase-type component at each iteration of the global EM.}

\item{initialpoint}{initial value for discretization of density}

\item{truncationpoint}{ultimate value for discretization of density}

\item{maxprobability}{max probability allowed for an interval in the discretization}

\item{maxdelta}{max size of interval allowed for the discretization}

\item{maxit}{maximum number of iterations when optimizing g function.}

\item{reltol}{relative tolerance when optimizing g function.}

\item{every}{number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{phasetype}.
}
\description{
Fit method for phase-type frailty models
}
\examples{
obj <- frailty(phasetype(structure = "general", dimension = 2), bhaz = "weibull", bhaz_pars = 2)
data <- sim(obj, n = 100)
fit(obj, data, stepsEM = 50, every = 10)
}
