% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{frailty}
\alias{frailty}
\title{Constructor function for univariate phase type frailty models}
\usage{
frailty(
  ph = NULL,
  bhaz = NULL,
  bhaz_pars = NULL,
  B = numeric(0),
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3
)
}
\arguments{
\item{ph}{An object of class \linkS4class{phasetype}.}

\item{bhaz}{Baseline hazard function.}

\item{bhaz_pars}{The parameters of the baseline hazard function.}

\item{B}{Regression parameters.}

\item{alpha}{A probability vector.}

\item{S}{A sub-intensity matrix.}

\item{structure}{A valid phase-type structure.}

\item{dimension}{The dimension of the phase-type structure (if provided).}
}
\value{
An object of class \linkS4class{frailty}.
}
\description{
Constructor function for univariate phase type frailty models
}
\examples{
frailty(phasetype(structure = "coxian", dimension = 4), bhaz = "weibull", bhaz_pars = 3)
}
