% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{frailty}
\alias{frailty}
\title{Constructor function for univariate phase type frailty models}
\usage{
frailty(
  ph = NULL,
  bhaz = NULL,
  bhaz_pars = NULL,
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3,
  B = numeric(0)
)
}
\arguments{
\item{ph}{an object of class \linkS4class{phasetype}.}

\item{bhaz}{baseline hazard function}

\item{bhaz_pars}{the parameters of the baseline hazard function}

\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid phase-type structure}

\item{dimension}{the dimension of the phase-type structure (if provided)}

\item{B}{regression parameters}
}
\value{
An object of class \linkS4class{frailty}.
}
\description{
Constructor function for univariate phase type frailty models
}
\examples{
frailty(phasetype(structure = "coxian", dimension = 4), bhaz = "weibull", bhaz_pars = 3)
}
