% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{haz,phasetype-method}
\alias{haz,phasetype-method}
\title{Hazard rate Method for phase type distributions}
\usage{
\S4method{haz}{phasetype}(x, y)
}
\arguments{
\item{x}{an object of class \linkS4class{phasetype}.}

\item{y}{a vector of locations.}
}
\value{
A list containing the locations and corresponding hazard rate evaluations.
}
\description{
Hazard rate Method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
haz(obj, c(1, 2, 3))
}
