% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{moment,phasetype-method}
\alias{moment,phasetype-method}
\title{Moment method for phase-type distributions}
\usage{
\S4method{moment}{phasetype}(x, k = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{phasetype}.}

\item{k}{A positive integer (moment order).}
}
\value{
The raw moment of the \linkS4class{phasetype} (or undelying \linkS4class{phasetype}) object.
}
\description{
Moment method for phase-type distributions
}
\examples{
set.seed(123)
obj <- phasetype(structure = "general", dimension = 3)
moment(obj, 2)
}
