% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_cdf}
\alias{ph_cdf}
\title{Phase-type cdf or tail}
\usage{
ph_cdf(x, alpha, S, lower_tail = TRUE)
}
\arguments{
\item{x}{Non-negative values.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{lower_tail}{Cdf or tail.}
}
\value{
The density at \code{x}.
}
\description{
Computes the cdf or tail of a phase-type distribution with parameters
\code{alpha} and \code{S} at \code{x}.
}
