% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_laplace_der_nocons}
\alias{ph_laplace_der_nocons}
\title{Derivative of order n of the Laplace transform of a phase-type distribution
without the multiplying constant}
\usage{
ph_laplace_der_nocons(r, n, alpha, S)
}
\arguments{
\item{r}{Real value.}

\item{n}{An integer.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}
}
\value{
Laplace transform at \code{r}.
}
\description{
Computes the derivative of order n (without the multiplying constant) of the
Laplace transform at \code{r} of a phase-type distribution with parameters
\code{alpha} and \code{S}.
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
S <- matrix(c(c(-1, 0, 0), c(1, -2, 0),c(0, 1, -5)), nrow = 3, ncol = 3)
ph_laplace_der_nocons(0.5, 2, alpha, S)
}
