% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{quan,phasetype-method}
\alias{quan,phasetype-method}
\title{Quantile Method for phase type distributions}
\usage{
\S4method{quan}{phasetype}(x, p)
}
\arguments{
\item{x}{an object of class \linkS4class{phasetype}.}

\item{p}{a vector of probabilities.}
}
\value{
A list containing the probabilities and corresponding quantile evaluations.
}
\description{
Quantile Method for phase type distributions
}
\examples{
obj <- phasetype(structure = "general")
quan(obj, c(0.5, 0.9, 0.99))
}
