% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\docType{class}
\name{shared-class}
\alias{shared-class}
\title{Shared phase type frailty model}
\value{
Class object
}
\description{
Class of objects for shared phase type frailty models
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution.}

\item{\code{bhaz1}}{a list comprising of the parameters.}

\item{\code{bhaz2}}{a list comprising of the parameters.}

\item{\code{coefs}}{regression parameters}
}}

