% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{show,phasetype-method}
\alias{show,phasetype-method}
\title{Show Method for phase type distributions}
\usage{
\S4method{show}{phasetype}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{phasetype}.}
}
\description{
Show Method for phase type distributions
}
