% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.phfrailty.R
\name{sim,frailty-method}
\alias{sim,frailty-method}
\title{Simulation Method for inhomogeneous phase type distributions}
\usage{
\S4method{sim}{frailty}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{frailty}.}

\item{n}{an integer of length of realization.}
}
\value{
A realization of independent and identically distributed phase-type frailty variables.
}
\description{
Simulation Method for inhomogeneous phase type distributions
}
\examples{
obj <- frailty(phasetype(structure = "general"), gfun = "lognormal", gfun_pars = 2)
sim(obj, n = 100)
}
