% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.sharedph.R
\name{surv,shared-method}
\alias{surv,shared-method}
\title{Survival method for shared phase-type frailty models}
\usage{
\S4method{surv}{shared}(x, q, X = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{shared}.}

\item{q}{A matrix of locations.}

\item{X}{A matrix of covariates.}
}
\value{
A vector containing the joint survival evaluations at the given locations.
}
\description{
Survival method for shared phase-type frailty models
}
\examples{
ph_obj <- phasetype(structure = "coxian")
shared_obj <- shared(ph_obj, bhaz1 = "weibull", bhaz_pars1 = 3, bhaz2 = "weibull", bhaz_pars2 = 2)
surv(shared_obj, matrix(c(1, 2), ncol = 2))
}
