% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{surv}
\alias{surv}
\title{New generic for the survival function of matrix distributions}
\usage{
surv(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
survival function from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{phasetype}
}
