% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices}
\alias{vector_of_matrices}
\title{Computes elements S^n until the value size_vec}
\usage{
vector_of_matrices(the_vector, S, size_vec)
}
\arguments{
\item{the_vector}{A vector to save results.}

\item{S}{Sub-intensity matrix.}

\item{size_vec}{Size of vector.}
}
\value{
Modified vector with the elements S^n.
}
\description{
Computes elements S^n until the value size_vec
}
